feat: tuning search with numpy candidate scorer
Random search over weights, pool size and relaxation geometry, scored on
validation pinball with refusals imputed at the naive-baseline pinball so
refusing can't game the objective. Term distances are polars-evaluated by
the serving expressions themselves; the numpy layer redoes only the
aggregation, and its parity with find_comps + summarize_pool is pinned by
synthetic tests plus a runtime gate on real validation queries inside the
tune stage. Parity work surfaced a real serving quirk - Float32-typed
terms (age, elo, minutes) multiplied by their weight in Float32 - fixed by
casting terms to Float64 before weighting (~1e-8 distance hygiene, no
ranking behavior change). numpy lands in a new eval dependency group,
synced in CI.

+++ server/app/services/comps.py
@@ -238,9 +238,12 @@ def _score(
         how="left",
     )
     terms = _distance_terms(query, dest_league, dest_club, drop_club_terms, config)
+    # Terms over Float32 source columns (age, elo, minutes) would otherwise
+    # multiply by the weight in Float32; cast so the whole distance is
+    # uniform double-precision arithmetic.
     numerator = reduce(
         operator.add,
-        [pl.when(d.is_not_null()).then(d * w).otherwise(0.0) for d, w in terms],
+        [pl.when(d.is_not_null()).then(d.cast(pl.Float64) * w).otherwise(0.0) for d, w in terms],
     )
     weight_mass = reduce(
         operator.add,
